refactor: move ORM models to dedicated package
