+++ pkgs/standards/peagen/peagen/orm/config/peagen_toml_spec.py
@@ -89,6 +89,3 @@ class PeagenTomlSpecModel(BaseModel):
             f"<PeagenTomlSpec id={self.id} tenant={self.tenant_id} "
             f"name={self.name!r} version={self.schema_version}>"
         )
-
-
-PeagenTomlSpec = PeagenTomlSpecModel
