+++ pkgs/standards/peagen/peagen/orm/security/__init__.py
@@ -1,3 +1,5 @@
 """Security-related ORM models."""
 
-from .public_key import PublicKey as PublicKey
+from .public_key import PublicKeyModel
+
+PublicKey = PublicKeyModel
