+++ pkgs/standards/peagen/peagen/orm/task/task_run.py
@@ -141,6 +141,3 @@ class TaskRunModel(BaseModel):
         tr.commit_hexsha = task.commit_hexsha
         tr.oids = task.oids
         return tr
-
-
-TaskRun = TaskRunModel
