+++ pkgs/standards/peagen/peagen/orm/tenant/tenant_user_association.py
@@ -66,6 +66,3 @@ class TenantUserAssociationModel(BaseModel):
             f"<TenantUserAssociation tenant_id={self.tenant_id} "
             f"user_id={self.user_id} role={self.role!r}>"
         )
-
-
-TenantUserAssociation = TenantUserAssociationModel
