+++ pkgs/standards/peagen/peagen/plugins/result_backends/in_memory_backend.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from peagen.orm.task.task_run import TaskRun
+from peagen.orm import TaskRun
 from .base import ResultBackendBase
