+++ pkgs/standards/peagen/tests/unit/test_result_backends.py
@@ -2,7 +2,7 @@ import json
 import uuid
 import pytest
 
-from peagen.orm.task.task_run import TaskRun
+from peagen.orm import TaskRun
 from peagen.orm.status import Status
 from peagen.plugins.result_backends.localfs_backend import LocalFsResultBackend
 from peagen.plugins.result_backends.postgres_backend import PostgresResultBackend
